changes for python bindings
