+++ scripts/build_all_static.sh
@@ -14,13 +14,14 @@ cp "$SDK_PACKAGE" "$TMP/unpack/"
 cd "$TMP/unpack"
 tar xvf spectacularAI_*.tar.gz
 make PREFIX="$TMP/install" test
+make install
 
 cd "$ROOT"
 ./scripts/build_all.sh -DspectacularAI_DIR="$TMP/install/lib/cmake/spectacularAI"
 
 INSTALL_DIR="spectacularai_ros2/install/spectacularai_ros2/"
 # Strip the executable (just in case the CMake build failed to do that)
-strip "$INSTALL_DIR/lib/libvislam.so"
+strip "$INSTALL_DIR/lib/libspectacularai_ros2.so"
 
 # Copy license/notice file form the SDK package
 LICENSE_OUTPUT_DIR="$INSTALL_DIR"
