+++ spectacularai_ros2/include/spectacularai_ros2/helpers.hpp
